pyproject.toml
[project]
name = "cinereserve-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Ailton-F",email = "ailtonxdz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-filter (>=25.2,<26.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "redis (>=7.3.0,<8.0.0)",
    "django-redis (>=6.0.0,<7.0.0)",
    "celery (>=5.6.2,<6.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]

[tool.poetry]
packages = [{include = "cinereserve-api", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = ">=9.0.2,<10.0.0"
pytest-django = ">=4.10.0,<5.0.0"
pytest-celery = ">=0.0.0"
